split getting started guide into a view and a form component
